Optimize QuadTree Insertion
Based on a tip by Robin Heggelund Hansen (of HAMT fame), this commit
inlines record updates. This gives slightly larger assets, but is worth
it in inner loops: There is roughly a 6X speed increase in the record
update alone, roughly 50% increase as a whole for insertBy

I also fixed a problem with the link that was caused by earlier versions
of elm-format (the latest version has fixed the issue).